requirements/packaginator.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Uncomment this if you want to support launchpad
#--requirement=launchpad.txt

django_extensions==0.4.1
simplejson==2.1.1
github2==0.2.0
south==0.7.2
django-tastypie==0.9.7
mimeparse==0.1.2
python-dateutil==1.5
django-reversion==1.3.1
django-sorting==0.1
coverage==3.4
django-coverage==1.1.1
django-flatblocks==0.5.0
feedparser==5.0.1
psycopg2==2.4
django-classy-tags==0.3.3.1
djangorestframework==0.1

# Celery dependencies:
celery==2.2.4
kombu==1.0.4
django-celery==2.2.4
django-kombu==0.9.2

# Sentry
django-sentry==1.6.9.1
django-paging==0.2.2
django-indexer==0.2.1
uuid==1.30
